backend/requirements.txt
Flask
flask-cors
tensorflow
opencv-python
pillow
gunicorn
